requirement.txt
scikit-learn==1.2.2
matplotlib==3.7.1
apache-airflow-providers-amazon==8.1.0